document-ocr-service/requirements.txt
pyenchant
python-Levenshtein
Levenshtein
numpy
pandas
opencv-python
matplotlib
pymorphy2
imutils
SQLAlchemy
psycopg2
psycopg2-binary
python-doctr
pytesseract
